Issue #649 Fixed

+++ src/main/java/org/powertac/officecomplexcustomer/customers/OfficeComplex.java
@@ -1743,7 +1743,8 @@ public class OfficeComplex extends AbstractCustomer
     if (tariff.getMinDuration() == 0)
       minDuration = OfficeComplexConstants.MEAN_TARIFF_DURATION;
     else
-      minDuration = tariff.getMinDuration() / TimeService.DAY;
+      minDuration =
+        (double) (tariff.getMinDuration()) / (double) (TimeService.DAY);
 
     log.debug("Minimum Duration: " + minDuration);
     return (lifecyclePayment / minDuration);
